+++ src/components/download/__tests__/DownloadButton.spec.js
@@ -6,22 +6,13 @@ import DownloadButton from '../DownloadButton.js'
 
 jest.mock('../../../actions/download', () => {
     return {
-        toggleDownloadDialog: jest
+        setDownloadMode: jest
             .fn()
-            .mockReturnValue({ type: 'DOWNLOAD_DIALOG_TOGGLE' }),
+            .mockReturnValue({ type: 'DOWNLOAD_MODE_SET' }),
     }
 })
 
 /* eslint-disable react/prop-types */
-jest.mock('../DownloadDialog.js', () => {
-    return {
-        __esModule: true,
-        default: function Mock(props) {
-            return <div className="ui-DownloadDialog">{props.children}</div>
-        },
-    }
-})
-
 jest.mock('../../core/index.js', () => {
     return {
         MenuButton: function Mock(props) {
@@ -34,7 +25,7 @@ jest.mock('../../core/index.js', () => {
 const mockStore = configureMockStore()
 
 describe('DownloadButton', () => {
-    it('renders menu button and download dialog', () => {
+    it('renders a download button', () => {
         const store = {}
 
         const { container } = render(