+++ src/components/map/MapView.js
@@ -22,6 +22,7 @@ const MapView = (props) => {
         closeCoordinatePopup,
         openContextMenu,
         setAggregations,
+        setMapObject,
         resizeCount,
     } = props
 
@@ -53,6 +54,7 @@ const MapView = (props) => {
                             interpretationModalOpen={interpretationModalOpen}
                             openContextMenu={openContextMenu}
                             resizeCount={resizeCount}
+                            setMapObject={setMapObject}
                         />
                     ) : (
                         <BaseUrlProvider>
@@ -69,6 +71,7 @@ const MapView = (props) => {
                                 openContextMenu={openContextMenu}
                                 setAggregations={setAggregations}
                                 resizeCount={resizeCount}
+                                setMapObject={setMapObject}
                             />
                         </BaseUrlProvider>
                     )}
@@ -92,6 +95,7 @@ MapView.propTypes = {
     openContextMenu: PropTypes.func,
     resizeCount: PropTypes.number,
     setAggregations: PropTypes.func,
+    setMapObject: PropTypes.func,
 }
 
 export default MapView