+++ src/components/orgunits/OrgUnitSelect.js
@@ -1,35 +1,15 @@
 import { OrgUnitDimension } from '@dhis2/analytics'
-import { useDataQuery } from '@dhis2/app-runtime'
 import { CenteredContent, CircularLoader, Help } from '@dhis2/ui'
 import cx from 'classnames'
 import PropTypes from 'prop-types'
-import React, { useCallback, useEffect } from 'react'
+import React, { useCallback } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { setOrgUnits } from '../../actions/layerEdit.js'
-import { DEFAULT_ORG_UNIT_LEVEL } from '../../constants/layers.js'
 import { translateOrgUnitLevels } from '../../util/orgUnits.js'
+import { useOrgUnits } from '../OrgUnitsProvider.js'
 import AssociatedGeometrySelect from './AssociatedGeometrySelect.js'
 import OrgUnitSelectMode from './OrgUnitSelectMode.js'
 import styles from './styles/OrgUnitSelect.module.css'
-
-// Fetches the root org units associated with the current user with fallback to data capture org units
-const ORG_UNIT_TREE_QUERY = {
-    tree: {
-        resource: 'organisationUnits',
-        params: () => ({
-            fields: ['id', 'displayName~rename(name)', 'path'],
-            userDataViewFallback: true,
-        }),
-    },
-    levels: {
-        resource: 'organisationUnitLevels',
-        params: {
-            fields: ['id', 'displayName~rename(name)', 'level'],
-            order: 'level:asc',
-            paging: false,
-        },
-    },
-}
 
 const TWO_OTHER_SELECTS = 'two'
 const ONE_OTHER_SELECT = 'one'
@@ -41,11 +21,9 @@ const OrgUnitSelect = ({
     hideSelectMode = true,
     hideLevelSelect = false,
     hideGroupSelect = false,
-    selectDefaultLevel = false,
-    selectRoots = false,
     warning,
 }) => {
-    const { loading, data, error } = useDataQuery(ORG_UNIT_TREE_QUERY)
+    const { roots, levels, loading, error } = useOrgUnits()
     const rows = useSelector((state) => state.layerEdit.rows)
     const dispatch = useDispatch()
 
@@ -60,31 +38,19 @@ const OrgUnitSelect = ({
         [dispatch]
     )
 
-    const roots = data?.tree.organisationUnits
-    const orgUnitLevels = data?.levels.organisationUnitLevels
-    const defaultLevel = orgUnitLevels?.[DEFAULT_ORG_UNIT_LEVEL]
-
     const orgUnits = translateOrgUnitLevels(
         rows?.find((r) => r.dimension === 'ou'),
-        orgUnitLevels
+        levels
     )
     const hasOrgUnits = !!orgUnits.length
 
-    useEffect(() => {
-        if (!rows && !hasOrgUnits && selectRoots && roots) {
-            setOrgUnitItems({
-                items: roots,
-            })
-        }
-    }, [rows, selectRoots, roots, hasOrgUnits, setOrgUnitItems])
-
-    useEffect(() => {
-        if (!rows && !hasOrgUnits && selectDefaultLevel && defaultLevel) {
-            setOrgUnitItems({ items: [{ id: `LEVEL-${defaultLevel.id}` }] })
-        }
-    }, [rows, selectDefaultLevel, defaultLevel, hasOrgUnits, setOrgUnitItems])
-
-    if (loading) {
+    if (error?.message) {
+        return (
+            <div className={styles.orgUnitSelect}>
+                <Help error>{error.message}</Help>
+            </div>
+        )
+    } else if (loading) {
         return (
             <div className={styles.loader}>
                 <CenteredContent>
@@ -92,8 +58,6 @@ const OrgUnitSelect = ({
                 </CenteredContent>
             </div>
         )
-    } else if (error?.message) {
-        return <Help error>{error.message}</Help>
     }
 
     const numOtherSelects =
@@ -134,8 +98,6 @@ OrgUnitSelect.propTypes = {
     hideLevelSelect: PropTypes.bool,
     hideSelectMode: PropTypes.bool,
     hideUserOrgUnits: PropTypes.bool,
-    selectDefaultLevel: PropTypes.bool,
-    selectRoots: PropTypes.bool,
     warning: PropTypes.string,
 }