+++ src/util/dataDownload.js
@@ -44,7 +44,13 @@ export const addPropNames = (layer, data) => {
           )
 }
 
-const standardizeFilename = (rawName) => rawName.replace(/\s+/g, '_')
+// Replaces anything that's not a letter or a number, and makes it all lower case
+// If no rawName the filename will start with "map_"
+// A short random string is added to avoid multiple files with the same filename
+export const standardizeFilename = (rawName, ext) =>
+    `${
+        rawName ? rawName.replace(/[^a-z0-9]/gi, '_').toLowerCase() : 'map_'
+    }_${Math.random().toString(36).substring(7)}.${ext}`
 
 export const createGeoJsonBlob = (data) => {
     const geojson = {
@@ -128,7 +134,7 @@ export const downloadData = async ({
 
     FileSaver.saveAs(
         createGeoJsonBlob(data),
-        standardizeFilename(name) + '.geojson'
+        standardizeFilename(name, 'geojson')
     )
 
     return true