+++ src/util/favorites.js
@@ -89,6 +89,7 @@ const models = ['program', 'programStage', 'organisationUnitGroupSet']
 const validModelProperties = [
     'id',
     'name',
+    'path',
     'dimensionType',
     'dimensionItemType',
 ]